+++ AppBase.Common/Enums/ImportOptions.cs
@@ -1,6 +1,0 @@
-namespace AppBase.Common.Enums;
-
-public enum ImportOptions
-{
-    OleDb, Odbc, xlsx, txt, netezza, db2, oracle, mssql, postgres/*,sqllite*/
-};